search_engine/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
elasticsearch==7.6.0
elasticsearch_dsl==7.1.0
gunicorn==20.0.4
django-jenkins==0.110.0
pandas==1.0.4
djangorestframework==3.11.1
nltk==3.4.5
numpy==1.17.0
spacy==2.0.12
wrapt==1.11.1
tensorflow==2.2.0
torch==1.6.0+cu101
transformers==4.0.0
html2text==2020.1.16
pyyaml==5.4.1 
uritemplate==3.0.1
sentence-transformers